# trabaio/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Sistema de Petshop",
    "version": "1.0.0",
    "description": "API para gerenciamento de petshop com autenticação JWT"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Listar todos os usuários",
        "tags": ["Usuários"],
        "security": [
          {
            "AdminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Usuário criado com sucesso"},
          "400": {"description": "Nome do usuário já existe"},
          "500": {"description": "Erro ao criar o usuário"}
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "tags": ["Usuários"],
        "security": [
          {"AdminAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {"description": "Usuário não encontrado"},
          "500": {"description": "Erro ao buscar usuário"}
        }
      },
      "put": {
        "summary": "Atualizar usuário por ID (completo)",
        "tags": ["Usuários"],
        "security": [
          {"AdminAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do usuário"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Usuário atualizado com sucesso"},
          "400": {"description": "Nenhum campo válido fornecido"},
          "404": {"description": "Usuário não encontrado ou dados idênticos"},
          "500": {"description": "Erro ao atualizar usuário"}
        }
      },
      "patch": {
        "summary": "Atualizar usuário por ID (parcial)",
        "tags": ["Usuários"],
        "security": [
          {"AdminAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do usuário"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdatePartial"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Usuário atualizado parcialmente com sucesso"},
          "400": {"description": "Nenhum campo fornecido"},
          "404": {"description": "Usuário não encontrado ou dados idênticos"},
          "500": {"description": "Erro ao atualizar usuário parcialmente"},
          "delete": {
            "summary": "Deletar usuário por ID",
            "tags": ["Usuários"],
            "security": [
              {"AdminAuth": []}
            ],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID do usuário"
              }
            ],
            "responses": {
              "200": {"description": "Usuário deletado com sucesso"},
              "404": {"description": "Usuário não encontrado"},
              "500": {"description": "Erro ao deletar usuário"}
            }
          }
        }
      }
    },
...
  }
}